+++ tests/Functional/Crawling/config-step-2.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Tests\Functional\Crawling;
 
-use Sobak\Scrawler\Block\FieldDefinition\StringField;
 use Sobak\Scrawler\Block\ResultWriter\InMemoryResultWriter;
 use Sobak\Scrawler\Block\UrlListProvider\ArgumentAdvancerUrlListProvider;
 use Sobak\Scrawler\Configuration\Configuration;
@@ -24,7 +23,7 @@ $scrawler
     ->addUrlListProvider(new ArgumentAdvancerUrlListProvider("{$host}/page-%u.html", 2, 2, 4))
     ->addObjectDefinition('message', new CssSelectorListMatcher('div.message'), function (ObjectConfiguration $object) {
         $object
-            ->addFieldDefinition('match', new StringField(new CssSelectorTextMatcher('span.content')))
+            ->addFieldDefinition('match', new CssSelectorTextMatcher('span.content'))
             ->addEntityMapping(SimpleMatchEntity::class)
             ->addResultWriter(SimpleMatchEntity::class, new InMemoryResultWriter())
         ;